Move custom ShaderWarmUp to PaintingBinding
