+++ dev/benchmarks/macrobenchmarks/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart' show DefaultShaderWarmUp;
 
 import 'common.dart';
 import 'src/cubic_bezier.dart';
@@ -7,35 +6,7 @@ import 'src/cull_opacity.dart';
 
 const String kMacrobenchmarks ='Macrobenchmarks';
 
-class CubicBezierShaderWarmUp extends DefaultShaderWarmUp {
-  @override
-  void warmUpOnCanvas(Canvas canvas) {
-    super.warmUpOnCanvas(canvas);
-
-    // Warm up the cubic shaders used by CubicBezierPage.
-    //
-    // This tests that our custom shader warm up is working properly.
-    // Without this custom shader warm up, the worst frame time is about 115ms.
-    // With this, the worst frame time is about 70ms. (Data collected on a Moto
-    // G4 based on Flutter version 704814c67a874077710524d30412337884bf0254.
-    final Path path = Path();
-    path.moveTo(20, 20);
-    // This cubic path is copied from
-    // https://skia.org/user/api/SkPath_Reference#SkPath_cubicTo
-    path.cubicTo(300, 80, -140, 90, 220, 10);
-    final Paint paint = Paint();
-    paint.isAntiAlias = true;
-    paint.strokeWidth = 18.0;
-    paint.style = PaintingStyle.stroke;
-    paint.strokeCap = StrokeCap.round;
-    canvas.drawPath(path, paint);
-  }
-}
-
-void main() => runApp(
-  MacrobenchmarksApp(),
-  shaderWarmUp: CubicBezierShaderWarmUp(),
-);
+void main() => runApp(MacrobenchmarksApp());
 
 class MacrobenchmarksApp extends StatelessWidget {
   @override
